move crypto/bn256 to erigon-lib
