+++ erigon-lib/crypto/bn256/bn256_slow.go
@@ -2,24 +2,22 @@
 // Use of this source code is governed by a BSD-style license that can be found
 // in the LICENSE file.
 
-//go:build amd64 || arm64
+//go:build !amd64 && !arm64
 
 // Package bn256 implements the Optimal Ate pairing over a 256-bit Barreto-Naehrig curve.
 package bn256
 
-import (
-	bn256cf "github.com/erigontech/erigon/crypto/bn256/cloudflare"
-)
+import bn256 "github.com/erigontech/erigon-lib/crypto/bn256/google"
 
 // G1 is an abstract cyclic group. The zero value is suitable for use as the
 // output of an operation, but cannot be used as an input.
-type G1 = bn256cf.G1
+type G1 = bn256.G1
 
 // G2 is an abstract cyclic group. The zero value is suitable for use as the
 // output of an operation, but cannot be used as an input.
-type G2 = bn256cf.G2
+type G2 = bn256.G2
 
 // PairingCheck calculates the Optimal Ate pairing for a set of points.
 func PairingCheck(a []*G1, b []*G2) bool {
-	return bn256cf.PairingCheck(a, b)
+	return bn256.PairingCheck(a, b)
 }
