+++ erigon-lib/crypto/bn256/cloudflare/lattice.go
