+++ erigon-lib/crypto/bn256/google/constants.go
