+++ tests/fuzzers/bn256/bn256_fuzz.go
@@ -14,8 +14,9 @@ import (
 	"math/big"
 
 	"github.com/consensys/gnark-crypto/ecc/bn254"
-	cloudflare "github.com/erigontech/erigon/crypto/bn256/cloudflare"
-	google "github.com/erigontech/erigon/crypto/bn256/google"
+
+	cloudflare "github.com/erigontech/erigon-lib/crypto/bn256/cloudflare"
+	google "github.com/erigontech/erigon-lib/crypto/bn256/google"
 )
 
 func getG1Points(input io.Reader) (*cloudflare.G1, *google.G1, *bn254.G1Affine) {
